Add files via upload
This program is used to calculate an employees net pay using their employee number. It ask the user to enter the gross pay, fed tax, state tax, and fica withholding for the employee number entered. It then does the math and displays the net pay, and also all the withholding fields like a paystub.